pyproject.toml
[tool.poetry]
name = "ort_testing"
version = "0.1.0"
description = ""
authors = ["1999DDm1999 <kazyevadina1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Pillow = "^9.1.1"
environs = "^9.5.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
